+++ src/components/Buttons/ShapeButton.tsx
@@ -1,33 +1,26 @@
 import React from 'react'
 import HomeIcon from '@heroicons/react/outline/HomeIcon'
-import { ShadowVariants } from '../../interfaces/types'
 import { borderRadius } from 'lib/shape'
 import { composeClasses } from 'lib/classes'
+import { ShadowVariants } from '../../interfaces/types'
 
 interface BaseCircleButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  Icon?: () => React.ReactNode
+  icon?: React.ReactNode
   disabled?: boolean
   selected?: boolean
   width?: string
-  height: string
+  height?: string
   shadow?: ShadowVariants
   variant?: 'square' | 'circle'
-  iconProperties?: { width?: string; height?: string; padding?: string }
 }
 
-export type WithoutTypeButtonProps = Omit<BaseCircleButtonProps, 'variant'>
+export type WithouVariantButtonProps = Omit<BaseCircleButtonProps, 'variant'>
 
 const globalStyles = {
   default: 'text-blue-700 hover:bg-gray-50',
   selected: 'bg-blue-700 text-white',
   disabled: 'text-gray-300'
-}
-
-const defaultIconProperties = {
-  width: '2.5rem',
-  height: '2.5rem',
-  padding: '0.5rem'
 }
 
 const buildClassName = (disabled?: boolean, selected?: boolean) => {
@@ -39,13 +32,12 @@ const buildClassName = (disabled?: boolean, selected?: boolean) => {
 }
 
 function BaseCircleButton({
-  Icon,
+  icon,
   width = '3rem',
   height = '3rem',
   disabled,
   selected,
   shadow = 'sm',
-  iconProperties = defaultIconProperties,
   variant = 'circle',
   ...props
 }: BaseCircleButtonProps) {
@@ -59,17 +51,26 @@ function BaseCircleButton({
         buildClassName(disabled, selected)
       )}
     >
-      {!Icon && <HomeIcon style={iconProperties} />}
-      {Icon && Icon()}
+      {icon ? (
+        icon
+      ) : (
+        <HomeIcon
+          style={{
+            width: '2.5rem',
+            height: '2.5rem',
+            padding: '0.5rem'
+          }}
+        />
+      )}
     </button>
   )
 }
 
-function SquareButton({ ...props }: WithoutTypeButtonProps) {
+function SquareButton({ ...props }: WithouVariantButtonProps) {
   return <BaseCircleButton variant="square" {...props} />
 }
 
-function CircleButton({ ...props }: WithoutTypeButtonProps) {
+function CircleButton({ ...props }: WithouVariantButtonProps) {
   return <BaseCircleButton variant="circle" {...props} />
 }